pyproject.toml
[tool.poetry]
name = "ek-chatbot"
version = "0.1.0"
description = ""
authors = ["Sven Duve <svenduve@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
asyncpg = "^0.29.0"
openai = "^1.12.0"
langchain = "^0.1.9"
pandas = "^2.2.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
langchain-community = "^0.0.24"
langchain-core = "^0.1.26"
langchain-openai = "^0.0.7"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"
pytest-asyncio = "^0.23.5"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
